particle movement
this is not perfect, but it works.
currently managing particle movement through the shader, the particles travel a distance along the tangent of the curve at their position and then restart. particles are given random progress to mask this.

The other solution for this would be to handle it in javascript, but this would affect performance.

next step, integrating this into the larger project.

yippee